refactor: partnership 패키지 내 클래스 Place 패키지로 이동
